+++ src/components/App/components/public/Connectors/Connector/ListView/Iterable/index.js
@@ -46,6 +46,17 @@ class Iterable extends EditorComponent {
   // }
 
 
+  renderMainView() {
+
+    const {
+      inEditMode,
+    } = this.getEditorContext();
+
+    return inEditMode ? super.renderMainView() : this.renderChildren();
+
+  }
+
+
   renderChildren() {
 
     const {